Criterion.Measurement

Description

Benchmark measurement code.

Documentation

initializeTime :: IO ()

Set up time measurement.

criterion measures time using OS-specific APIs whenever possible for efficiency. On certain operating systems, such as macOS and Windows, one must explicitly initialize a timer (which initializeTime accomplishes) before one can actually measure the current time (which getTime accomplishes).

It is imperative that you call initializeTime before calling getTime. (See this bug report for an example of what can happen if you do not do so.) All of the IO-returning functions in Criterion.Main make sure that this is done, but other functions (such as those in Criterion.Measurement) do not guarantee this unless otherwise stated.

getTime :: IO Double

Return the current wallclock time, in seconds since some arbitrary time.

You must call initializeTime once before calling this function! Refer to the documentation for initializeTime for more details.

getCPUTime :: IO Double

Return the amount of elapsed CPU time, combining user and kernel (system) time into a single measure.

getCycles :: IO Word64

Read the CPU cycle counter.

getGCStatistics :: IO (Maybe GCStatistics)

Try to get GC statistics, bearing in mind that the GHC runtime will throw an exception if statistics collection was not enabled using "+RTS -T".

If you need guaranteed up-to-date stats, call performGC first.

data GCStatistics

Statistics about memory usage and the garbage collector. Apart from gcStatsCurrentBytesUsed and gcStatsCurrentBytesSlop all are cumulative values since the program started.

GCStatistics is cargo-culted from the GCStats data type that GHC.Stats used to export. Since GCStats was removed in GHC 8.4, criterion uses GCStatistics to provide a backwards-compatible view of GC statistics.

Constructors

GCStatistics

Fields

gcStatsBytesAllocated :: !Int64

Total number of bytes allocated

gcStatsNumGcs :: !Int64

Number of garbage collections performed (any generation, major and minor)

gcStatsMaxBytesUsed :: !Int64

Maximum number of live bytes seen so far

gcStatsNumByteUsageSamples :: !Int64

Number of byte usage samples taken, or equivalently the number of major GCs performed.

gcStatsCumulativeBytesUsed :: !Int64

Sum of all byte usage samples, can be used with gcStatsNumByteUsageSamples to calculate averages with arbitrary weighting (if you are sampling this record multiple times).

gcStatsBytesCopied :: !Int64

Number of bytes copied during GC

gcStatsCurrentBytesUsed :: !Int64

Number of live bytes at the end of the last major GC

gcStatsCurrentBytesSlop :: !Int64

Current number of bytes lost to slop

gcStatsMaxBytesSlop :: !Int64

Maximum number of bytes lost to slop at any one time so far

gcStatsPeakMegabytesAllocated :: !Int64

Maximum number of megabytes allocated

gcStatsMutatorCpuSeconds :: !Double

CPU time spent running mutator threads. This does not include any profiling overhead or initialization.

gcStatsMutatorWallSeconds :: !Double

Wall clock time spent running mutator threads. This does not include initialization.

gcStatsGcCpuSeconds :: !Double

CPU time spent running GC

gcStatsGcWallSeconds :: !Double

Wall clock time spent running GC

gcStatsCpuSeconds :: !Double

Total CPU time elapsed since program start

gcStatsWallSeconds :: !Double

Total wall clock time elapsed since start

secs :: Double -> String

Convert a number of seconds to a string. The string will consist of four decimal places, followed by a short description of the time units.

measure

Arguments

:: Benchmarkable
Operation to benchmark.

-> Int64
Number of iterations.

-> IO (Measured, Double)

Measure the execution of a benchmark a given number of times.

This function initializes the timer before measuring time (refer to the documentation for initializeTime for more details).

runBenchmark

Arguments

:: Benchmarkable

-> Double
Lower bound on how long the benchmarking process should take. In practice, this time limit may be exceeded in order to generate enough data to perform meaningful statistical analyses.

-> IO (Vector Measured, Double)

Run a single benchmark, and return measurements collected while executing it, along with the amount of time the measurement process took.

This function initializes the timer before measuring time (refer to the documentation for initializeTime for more details).

runBenchmarkable :: Benchmarkable -> Int64 -> (a -> a -> a) -> (Int64 -> IO () -> IO a) -> IO a

runBenchmarkable_ :: Benchmarkable -> Int64 -> IO ()

measured :: Measured

An empty structure.

applyGCStatistics

Arguments

:: Maybe GCStatistics
Statistics gathered at the end of a run, post-GC.

-> Maybe GCStatistics
Statistics gathered at the end of a run, pre-GC.

-> Maybe GCStatistics
Statistics gathered at the beginning of a run.

-> Measured
Value to "modify".

-> Measured

Apply the difference between two sets of GC statistics to a measurement.

threshold :: Double

The amount of time a benchmark must run for in order for us to have some trust in the raw measurement.

We set this threshold so that we can generate enough data to later perform meaningful statistical analyses.

The threshold is 30 milliseconds. One use of runBenchmark must accumulate more than 300 milliseconds of total measurements above this threshold before it will finish.