Criterion.Measurement.Types.Internal

Description

Exports fakeEnvironment.

Documentation

fakeEnvironment :: env

A dummy environment that is passed to functions that create benchmarks from environments when no concrete environment is available.

nf' :: (b -> ()) -> (a -> b) -> a -> Int64 -> IO ()

Generate a function which applies an argument to a function a given number of times, reducing the result to normal form.

whnf' :: (a -> b) -> a -> Int64 -> IO ()

Generate a function which applies an argument to a function a given number of times.